create a components folder and move display posts from views
