+++ frontend/src/app/components/display-post/display-post.component.ts
