+++ frontend/src/app/views/post/post.component.html
@@ -1,0 +1,1 @@
+<p>post works!</p>
